Split non-display parts of image display into ImageDisplay.py.
